requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
timm>=0.9.2
spacy>=3.5.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
Pillow>=9.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
